dolt_fetch: Allow `dolt fetch` to work with username/password auth against a remotesapi.
When we migrated `dolt fetch` to use the SQL interface, we broke
username/password authentication for it. This change restores that
functionality.

Because `dolt fetch` can run against a SQL server, the semantics here are a bit
subtle. We pass along the `--user` parameter which is given to the `dolt fetch`
invocation. When the server sees a `--user` parameter in a `call dolt_fetch()`,
it consults its own process environment to pull out the password. In the common
case where `dolt fetch` is running against a SQL engine which is in the same
process, this means the `dolt fetch` invocations same the same.